+++ scripts/publish.sh
@@ -1,8 +1,4 @@
 #!/bin/bash
 
-if [ ! -f "./package.json" ]; then
-  echo "🛠  Generating package.json from base manifest..."
-  node scripts/manifest/build-manifest.js
-fi
-
+sh scripts/build.sh
 npm run publish