+++ src/openers/index.ts
@@ -2,6 +2,7 @@ export * as bot from "./bot/helpers";
 export * as dashboard from "./dashboard/helpers";
 export * as flexipage from "./flexipage/helpers";
 export * as flow from "./flow/helpers";
+export * as permissionset from "./permissionset/helpers";
 export * as report from "./report/helpers";
 export * as sobject from "./sobject/helpers";
 export * as validationRule from "./validationRule/helpers";